refactor: Add context manager for database
Replace  the connect() function with a connect_to_db()
function that is a context manager. This allows the "with as"
pattern to be used when interfacing with database, in a similar
way it is recommended when doing file IO in Python.

This then removes the need to explicitly close the database
connection in the code. This is done in the "finally:" block
of the new connect_to_db() function.