fix: Extract access token from JSON response
Since no version was given in the URL endpoint that
upgrades the short-lived token for the long-lived
token, it defaulted to the oldest version available.

This caused the code to break on 25th March 2017
as the old version became version 2.3 of the Facebook
Graph API, as a change in that version meant that the
access token was returned in JSON format. The code
change then extracts the token properly from the
response.